+++ R/best_reconstruction_QP.R
@@ -1,35 +1,67 @@
 #' Find "best" reconstruction of a target signature or spectrum from a set of signatures.
 #'
-#' @param target.sig The signature to reconstruct.
+#' @param target.sig The signature or spectrum to reconstruct;
+#'   a non-negative numeric vector or 1-column matrix-like object.
 #'
 #' @param sig.universe The universe of signatures from which to reconstruct
-#'   \code{target.sig}.
+#'   \code{target.sig}; a non-negative numeric matrix-like object with
+#'   \code{nrow(sig.universe) == length(target.sig)}. The sums of each column
+#'   must be 1. Must not contain duplicate columns or have other
+#'   non-unique quadratic programming solutions (not checked, but will generate
+#'   an error from `\link[quadprog]{solve.QP}` in package `quadprog`).
 #'
 #' @param max.subset.size Maximum number of signatures to use to
 #'   reconstruct \code{target.sig}.
 #'
-#' @param method As in \code{\link[philentropy]{dist_one_one}}, and used only
+#' @param method As in \code{\link[philentropy]{dist_one_one}} in
+#'   package \code{philentropy}, and used only
 #'   to find the final "best" reconstruction. The optimized exposures from which
 #'   to selected the "best" reconstruction are calculated using
-#'   \code{\link{optimize_exposure_QP}}, which uses \code{\link[quadprog]{solve.QP}}.
+#'   \code{\link{optimize_exposure_QP}}, which uses \code{\link[quadprog]{solve.QP}}
+#'   in package \code{quadprog}.
 #'
 #' @param trim.less.than After optimizing exposures with
 #'   \code{\link{optimize_exposure_QP}}, remove exposures less than
 #'   \code{trim.less.than} and then re-optimize.
 #'
-#' @return A list .... TODO
+#' @md
+#'
+#' @return A list with elements:
+#'
+#' * \code{optimized.exposure} A numerical vector of the exposures that
+#'   give the "best" reconstruction. This vector is empty if there is
+#'   an error.
+#'
+#' * \code{similarity} The similarity between the \code{reconstruction}
+#'   (see below) and \code{target.sig} according to the distance
+#'   or similarity provided by the \code{method} argument.
+#'
+#' * \code{method} The value specified for the `method` argument,
+#'   or an error message if `optimize.exposure` is empty.
+#'
+#' * \code{reconstruction} The reconstruction of `target.sig` according to
+#'   `optimized.exposure`.
 #'
 #' @export
 #'
 #' @details
-#'   Should normally be fast if you do not specify \code{max.subset.size},
-#'   but will be combinatorially slow if \code{max.subset.size} is large
+#'   This function should be fast if you do not specify \code{max.subset.size},
+#'   but it will be combinatorially slow if \code{max.subset.size} is large
 #'   and \code{trim.less.than} is small or negative. So do not do that.
-#'   TODO explain algorithm. TODO: tests
+#'   If `max.subset.size` is `NULL`, then the function just uses `\link{optimize_exposure_QP}`.
+#'   and then excludes exposures < `trim.less.than`, and then re-runs
+#'   `\link{optimize_exposure_QP}`. Otherwise, after excluding
+#'   exposures < `trim.less.than`, the `optimize_exposure_QP` on each
+#'   subset of signatures of size $le$ `max.subset.size`, removes exposures < `trim.less.than`,
+#'   reruns `\link{optimize_exposure_Q}`, calculates the reconstruction and
+#'   similarity between the reconstruction and the `target.sig` and returns the information for
+#'   the exposures that have the greatest similarity.
 #'
 #' @examples
 #' set.seed(888)
-#' sig.u <- do.call(cbind, lapply(1:6, function(x) runif(n = 96)))
+#'   sig.u <-
+#' do.call(cbind,
+#'        lapply(1:6, function(x) { col <- runif(n = 96); col/sum(col) }))
 #' rr <- best_reconstruction_QP(target.sig = sig.u[ , 1, drop = FALSE],
 #'                             sig.universe = sig.u[ , 2:6])
 #' names(rr)
@@ -47,8 +79,13 @@ best_reconstruction_QP <- function(target.sig,
                                    method          = "cosine",
                                    trim.less.than  = 1e-10) {
 
+  if (!isTRUE(all.equal(colSums(sig.universe), rep(1, ncol(sig.universe))))) {
+    return(list(optimized.exposure = numeric(),
+           method = "Error: all signature.universe columns must sum to 0"))
+  }
+
   if (is.null(colnames(sig.universe))) {
-    colnames(sig.universe) <- paste("v", 1:ncol(sig.universe))
+    colnames(sig.universe) <- paste("v", 1:ncol(sig.universe), sep = "")
   }
 
   optimize.etc <- function(target.sig, sig.u) {
@@ -102,11 +139,4 @@ best_reconstruction_QP <- function(target.sig,
   similarities <- unlist(lapply(rr, `[[`, "similarity"))
   best <- which(similarities == max(similarities))
   return(rr[[best]])
-}
-
-if (FALSE) {
-  s96 <- cosmicsig::COSMIC_v3.2$signature$GRCh37$SBS96
-  .tar <- s96[ , "SBS29", drop = F]
-  .u2 <- s96[ , -which(colnames(s96) == "SBS29"), drop = FALSE]
-  mSigTools:::best_reconstruction_QP(.tar, sig.universe = .u2)
 }